requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Frontend
streamlit>=1.35.0

# HTTP client (frontend -> backend)
httpx>=0.27.0

# Config
pyyaml>=6.0

# Image processing
Pillow>=10.3.0
numpy>=1.24.0

# ML — NOTE for Jetson Orin:
#   Install torch/torchvision from NVIDIA's Jetson index, NOT from PyPI.
#   See: https://forums.developer.nvidia.com/t/pytorch-for-jetson/72048
#   Example: pip install torch torchvision --index-url <jetson-wheel-url>
torch>=2.1.0
torchvision>=0.16.0

# HuggingFace
transformers>=4.40.0
diffusers>=0.27.0
accelerate>=0.29.0
huggingface_hub>=0.23.0
safetensors>=0.4.0
